12.1.4
Got SQL up and running.  We created a database using this command

CREATE DATABASE election;
then we selected it with this

USE election;

then we did this.

CREATE TABLE candidates (
ceates a table and names it candidates and use an open parens.

  id INTEGER AUTO_INCREMENT PRIMARY KEY,
then set the column name to id declaring it needs a number, primary key that each value in this column must be unique and auto_increment increments with each row and assigns a new value to the id.

  first_name VARCHAR(30) NOT NULL,
  last_name VARCHAR(30) NOT NULL,
both name  their columns first_name or last_name and then gives you 30 spaces to type in your name with variable character.  Not null says that there must be at least one character in this spot.

  industry_connected BOOLEAN NOT NULL
this is looking for a false(0) or true(1), if null it will not accept.
);

I coped pasted most of this in.  I would like to try typing to see how multiple lines work.  I assume it will just go to the next line until I kill it with ;.