12.3.7
created a put for party_id in the candidates table.  This allows us to update the candidates party.  put = edit.

app.put("/api/candidate/:id", (req, res) => {
  const errors = inputCheck(req.body, "party_id");

  if (errors) {
    res.status(400).json({ error: errors });
    return;
  }

  const sql = `UPDATE candidates SET party_id = ?
               WHERE id = ?`;
  const params = [req.body.party_id, req.params.id];
  db.query(sql, params, (err, result) => {
    if (err) {
      res.status(400).json({ error: err.message });
      // check if a record was found
    } else if (!result.affectedRows) {
      res.json({
        message: "Candidate not found",
      });
    } else {
      res.json({
        message: "success",
        data: req.body,
        changes: result.affectedRows,
      });
    }
  });
});

This code is a bit more involved.  But we create a check that tests if you selected a valid party id.

Then the next to lines are the Update which lays the ground work SET making sure the first value is the party_id from the body of what is being sent in at the location of the param from the URL that zeroes in on the candidates table row.  Then in the params var we make sure to call it correctly with req.body.part_id being the new id sent in and the req.params.id being the candidate being targeted by the URL.