12.5.4
Built the post to add a voter.  I again was able to copy and paste the candidatesRoutes.js and by carefully following the steps make it work before the module showed me.  No real new info here.  But it is worth noting we have not built the router.get so in order to see if it worked I had to go into the MySWL shell USE election;- SELECT * FROM votes; - to see that it worked after using insomnia to perform my post.